# ARTEMIS-1883 Improve FQQN validation