@rtista 2020-12-01
> can you squash these commits and push -f?

Yes of course! I am currently taking care of it ;)