@jbertram 2020-12-02
@rtista, are you still working on this or should I updated it and send a PR of my own?