@rtista 2020-12-03
> @rtista, are you still working on this or should I updated it and send a PR of my own?

Hey @jbertram , I've had a difficult day today but I'll make it up tomorrow! Sorry about the delay!