@clebertsuconic 2020-12-08
@rtista I will run the whole test suite and if it's passing it's a good sign.. takes 2 hours to run it.