@clebertsuconic 2020-12-08
@rtista I actually had 3 failures, although I don't know if the BasicSecurityTest is an intermittent / non related failure yet.

but the AmqpFullyQualifiedNameTest is definitely a real failure.

st Name
Duration
Age
 org.apache.activemq.artemis.tests.integration.amqp.AmqpFullyQualifiedNameTest.testQueueSpecial	79 ms	1
 org.apache.activemq.artemis.tests.integration.security.BasicSecurityManagerFailoverTest.testFailover[replicated=true]	0.48 sec	1
 org.apache.activemq.artemis.tests.integration.security.BasicSecurityManagerFailoverTest.testFailover[replicated=false]	0.27 sec	1