@rtista 2020-12-09
Thanks for the quick response @clebertsuconic, I was attempting to execute that same test in Intellij but I'm having what I believe is a different failure from yours:
 
`[main] 00:04:44,964 ERROR [org.apache.activemq.artemis.core.client] AMQ214032: Unable to initialize VersionLoader : java.lang.NumberFormatException: For input string: "${activemq.version.majorVersion}"`

I am probably missing some setup in order to run the test through Intellij.

I can see however that the test is attempting to create an invalid FQQN (since an empty address is being passed) and expecting it to fail as in the test doc says "Broker should return exception if no address is passed in FQQN".