@clebertsuconic 2020-12-09
I see that happening in Intelij sometimes.  

I don't know what magic I do.  But I mix refreshing the workspace with the proper JDK.  Pay attention to the profiles you have on. If you are on jdk8 only select jdk8 as the profile.