@rtista 2020-12-09
I just got it working by re-running install, maybe I missed something. Anyway, I already have that test failing with the expected exception not being thrown. I'll be chasing it tomorrow after work! Thanks again for the help! ;)