@clebertsuconic 2020-12-14
The test is trying to open a queue with ::queue.. after your change this is being considered a regular queue, which is not what it should be doing.

it seems the opposite of what you're changing.