@clebertsuconic 2020-12-14
To do what you're proposing, you would need some pedantic mode where you would throw an exception with something invalidException. however this type of semantic change may break other things..

Do you really want this PR? I think we should consider leaving it as is.