@rtista 2020-12-14
@clebertsuconic I fully understand your concerns, I could also just check for presence of characters on the right side of the separator. This would strengthen the validation by always checking for the queue part at least but ignoring the address which would still allow that test to succesfuly pass.

I guess I'd like to hear more opinions though, such as @jbertram's, before closing the PR as something that would cause more harm than good 😜

Regarding the PR, I do not "want" the PR, I was just looking to contribute. I picked up this Jira issue as it seemed like a good first issue in a large project such as this one. If you can direct me to other good Jira's for starting up artemis contributions I'd be delighted x)