@jbertram 2020-12-16
In principal, I liked the idea of tightening the rules for FQQN validation, but it's obviously causing other issues. Perhaps the method should throw an exception if the input is not a valid FQQN but it still contains the separator (i.e. `::`). That may cover all the necessary cases, but it could cause more problems as well. I'm not sure at this point. In any event, I'm happy to merge the PR if there are no test failures, but as it stands now it can't be merged.